app/requirements.txt
flask
markdown 

grpcio
grpcio-tools
googleapis-common-protos